requirements.txt
Flask==2.3.1
Flask-SQLAlchemy>=3.0.3
Flask-HTTPAuth>=4.1.0
SQLAlchemy>=1.3.17
urllib3>=2.0.0
Werkzeug>=2.3.0
flask-cors>=3.0.9
flask-accept>=0.0.6
python-dateutil>=2.8.1
python-editor>=1.0.4
coverage>=5.1.0
pydantic>=1.6.2
mimesis>=4.0.0
black>=23.3.0
Jinja2>=3.1.2
psycopg2-binary>=2.8.5
Flask-Cors>=3.0.10
types-flask-cors>=3.0.15
Werkzeug>=2.3.3
PyJWT>=2.7.0

